add new models & code led control algo

+++ tup-controller/src/reg/reg.cpp
@@ -14,21 +14,27 @@ void latch_enable()
     digitalWrite(LATCH_PIN, LOW);
 }
 
-void send_data(unsigned int data_out)
+void send_data(unsigned long data_out)
 {
     int i;
+    // Serial.println("F- ");
     for (i = 0; i < TOTAL_NUM_BITS; i++)
-    {
+    {   
+        // Serial.print(data_out);
+        // Serial.print(data_out >> i);
         if ((data_out >> i) & (0x01))
         {
             digitalWrite(DATA_PIN, HIGH);
+            // Serial.print(1);
         }
         else
         {
             digitalWrite(DATA_PIN, LOW);
+            // Serial.print(0);
         }
 
         clock_signal();
     }
+    // Serial.println(" <-");
     latch_enable();
 }
